Refactor: Move Pattern into pattern module
